Update player.cpp
Implemented a functional player.cpp file, player now moves left and right and has a display function with keyboard callback. Player is also smaller and position is centered in the middle of the screen along the x-axis. It's y-axis is stationary. Additionally, the player's width now cannot go beyond the screen and only its vertex touches the edges.